Made SpinnerView more testable
